% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePerm.R
\name{makePerm}
\alias{makePerm}
\title{Make data point permutations.}
\usage{
makePerm(N)
}
\arguments{
\item{N}{length of datasets}
}
\value{
indices of permuted dataset
}
\description{
THe function is extracted to be able to apply group permutations later.
}
